Move response struct and related code to separate file
