+++ js/modules/k6/http/response.go
@@ -1,0 +1,141 @@
+/*
+ *
+ * k6 - a next-generation load testing tool
+ * Copyright (C) 2016 Load Impact
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package http
+
+import (
+	"context"
+	"crypto/tls"
+	"encoding/json"
+
+	"github.com/dop251/goja"
+	"github.com/loadimpact/k6/js/common"
+	"github.com/loadimpact/k6/js/modules/k6/html"
+	"github.com/loadimpact/k6/lib"
+	"golang.org/x/crypto/ocsp"
+)
+
+type OCSP struct {
+	ProducedAt, ThisUpdate, NextUpdate, RevokedAt int64
+	RevocationReason                              string
+	Status                                        string
+}
+
+type HTTPResponseTimings struct {
+	Duration, Blocked, LookingUp, Connecting, Sending, Waiting, Receiving float64
+}
+
+type HTTPResponse struct {
+	ctx context.Context
+
+	RemoteIP       string
+	RemotePort     int
+	URL            string
+	Status         int
+	Proto          string
+	Headers        map[string]string
+	Body           string
+	Timings        HTTPResponseTimings
+	TLSVersion     string
+	TLSCipherSuite string
+	OCSP           OCSP `js:"ocsp"`
+	Error          string
+
+	cachedJSON goja.Value
+}
+
+func (res *HTTPResponse) setTLSInfo(tlsState *tls.ConnectionState) {
+	switch tlsState.Version {
+	case tls.VersionSSL30:
+		res.TLSVersion = SSL_3_0
+	case tls.VersionTLS10:
+		res.TLSVersion = TLS_1_0
+	case tls.VersionTLS11:
+		res.TLSVersion = TLS_1_1
+	case tls.VersionTLS12:
+		res.TLSVersion = TLS_1_2
+	}
+
+	res.TLSCipherSuite = lib.SupportedTLSCipherSuitesToString[tlsState.CipherSuite]
+	ocspStapledRes := OCSP{Status: OCSP_STATUS_UNKNOWN}
+
+	if ocspRes, err := ocsp.ParseResponse(tlsState.OCSPResponse, nil); err == nil {
+		switch ocspRes.Status {
+		case ocsp.Good:
+			ocspStapledRes.Status = OCSP_STATUS_GOOD
+		case ocsp.Revoked:
+			ocspStapledRes.Status = OCSP_STATUS_REVOKED
+		case ocsp.ServerFailed:
+			ocspStapledRes.Status = OCSP_STATUS_SERVER_FAILED
+		case ocsp.Unknown:
+			ocspStapledRes.Status = OCSP_STATUS_UNKNOWN
+		}
+		switch ocspRes.RevocationReason {
+		case ocsp.Unspecified:
+			ocspStapledRes.RevocationReason = OCSP_REASON_UNSPECIFIED
+		case ocsp.KeyCompromise:
+			ocspStapledRes.RevocationReason = OCSP_REASON_KEY_COMPROMISE
+		case ocsp.CACompromise:
+			ocspStapledRes.RevocationReason = OCSP_REASON_CA_COMPROMISE
+		case ocsp.AffiliationChanged:
+			ocspStapledRes.RevocationReason = OCSP_REASON_AFFILIATION_CHANGED
+		case ocsp.Superseded:
+			ocspStapledRes.RevocationReason = OCSP_REASON_SUPERSEDED
+		case ocsp.CessationOfOperation:
+			ocspStapledRes.RevocationReason = OCSP_REASON_CESSATION_OF_OPERATION
+		case ocsp.CertificateHold:
+			ocspStapledRes.RevocationReason = OCSP_REASON_CERTIFICATE_HOLD
+		case ocsp.RemoveFromCRL:
+			ocspStapledRes.RevocationReason = OCSP_REASON_REMOVE_FROM_CRL
+		case ocsp.PrivilegeWithdrawn:
+			ocspStapledRes.RevocationReason = OCSP_REASON_PRIVILEGE_WITHDRAWN
+		case ocsp.AACompromise:
+			ocspStapledRes.RevocationReason = OCSP_REASON_AA_COMPROMISE
+		}
+		ocspStapledRes.ProducedAt = ocspRes.ProducedAt.Unix()
+		ocspStapledRes.ThisUpdate = ocspRes.ThisUpdate.Unix()
+		ocspStapledRes.NextUpdate = ocspRes.NextUpdate.Unix()
+		ocspStapledRes.RevokedAt = ocspRes.RevokedAt.Unix()
+	}
+
+	res.OCSP = ocspStapledRes
+}
+
+func (res *HTTPResponse) Json() goja.Value {
+	if res.cachedJSON == nil {
+		var v interface{}
+		if err := json.Unmarshal([]byte(res.Body), &v); err != nil {
+			common.Throw(common.GetRuntime(res.ctx), err)
+		}
+		res.cachedJSON = common.GetRuntime(res.ctx).ToValue(v)
+	}
+	return res.cachedJSON
+}
+
+func (res *HTTPResponse) Html(selector ...string) html.Selection {
+	sel, err := html.HTML{}.ParseHTML(res.ctx, res.Body)
+	if err != nil {
+		common.Throw(common.GetRuntime(res.ctx), err)
+	}
+	if len(selector) > 0 {
+		sel = sel.Find(selector[0])
+	}
+	return sel
+}
